circuits_lib::bridge_circuit

Function deposit_constant

pub fn deposit_constant(
    operator_xonlypk: [u8; 32],
    watchtower_challenge_connector_start_idx: u32,
    watchtower_pubkeys: &[[u8; 32]],
    move_txid: [u8; 32],
    round_txid: [u8; 32],
    kickoff_round_vout: u32,
    genesis_state_hash: [u8; 32],
) -> DepositConstant

Computes a deposit constant hash using various transaction and cryptographic components.

Parameters

operator_xonlypk: A 32-byte array representing the operator’s X-only public key.

watchtower_challenge_connector_start_idx: A 16-bit unsigned integer marking the start index of the watchtower challenge connector.

watchtower_pubkeys: A slice of 32-byte arrays representing tweaked watchtower public keys.

move_txid: A 32-byte array representing the transaction ID of the move transaction.

round_txid: A 32-byte array representing the transaction ID of the round transaction.

kickoff_round_vout: A 32-bit unsigned integer indicating the vout of the kickoff round transaction.

genesis_state_hash: A 32-byte array representing the genesis state hash.

Returns

A DepositConstant containing a 32-byte SHA-256 hash of the concatenated input components.