circuits_lib::bridge_circuit

Module groth16_verifier

Bridge Circuit Groth16 Verifier

This module implements the Groth16 verifier for the bridge circuit. It includes functions to create digests for the work-only circuit output, verify Groth16 proofs, and handle the conversion of hexadecimal strings to decimal. The verifier uses the prepared verification keys and constants defined in the constants.rs module.

Key Functions

create_journal_digest: Creates a digest for the journal of the work-only circuit output.

create_output_digest: Creates an output digest for the work-only circuit output.

create_claim_digest: Creates a claim digest for the work-only circuit output.

CircuitGroth16WithTotalWork: A struct that encapsulates the Groth16 proof with total work and genesis state hash.

verify: Verifies the Groth16 proof against the prepared verification key and public inputs.

to_decimal: Converts a hexadecimal string to a decimal string representation.

Structs

CircuitGroth16WithTotalWork

Groth16 proof with total work and genesis state hash. In Clementine, this is provided by the watchtowers who challenge the operator whom they suspect of malicious behavior. Just by knowing the Groth16 proof and the total work, we can reconstruct the public outputs of the proof and verify it against the Verifying Key (VK) of the Groth16 proof.

Functions

create_claim_digest

Creates a claim digest for the work-only circuit output.

create_journal_digest

Creates a digest for the journal of the work-only circuit output.

create_output_digest

Creates an output digest for the work-only circuit output.

to_decimal

Converts a hexadecimal string to a decimal string representation.