circuits_lib::bridge_circuit
Module lc_proof

Light Client Proof Verifier

This module implements the light client proof verifier for the bridge circuit. It includes functions to verify light client proofs and extracting the light client circuit output.

Constants

LC_IMAGE_ID

Functions

check_method_id

deserialize_circuit_output 🔒

Deserializes the light client circuit output from journal bytes.

is_regtest 🔒

lc_proof_verifier

Verifies the light client proof and returns the light client circuit output.