circuits_lib::bridge_circuit
Module transaction

Bitcoin Transaction Wrapper

This module provides a wrapper around Bitcoin’s Transaction type, adding functionality for calculating transaction IDs, serializing and deserializing transactions, and providing a more convenient interface for working with Bitcoin transactions in the context of the bridge circuit.

Structs

CircuitTransaction

A wrapper around Bitcoin’s Transaction type that provides additional and simpler functionality.

Functions

deserialize_txin 🔒

deserialize_txout 🔒

serialize_txin 🔒

serialize_txout 🔒