circuits_lib::bridge_circuit::transaction
Struct CircuitTransaction

pub struct CircuitTransaction(pub Transaction);

Tuple Fields
0: Transaction
Implementations

impl CircuitTransaction

pub fn from(transaction: Transaction) -> Self

pub fn inner(&self) -> &Transaction

pub fn txid(&self) -> [u8; 32]

Returns the transaction id, in big-endian byte order. One must be careful when dealing with Bitcoin transaction ids, as they are little-endian in the Bitcoin protocol.

pub fn mid_state_txid(&self) -> [u8; 32]

Returns the first digest of the transaction to be used in SPV