circuits_lib::common

Module constants

Common Constants

Common constants used for Risc0 circuits in the Clementine protocol. These constants are used across various modules to ensure consistency and correctness in the circuit operations. They include constants for the number of outputs, transaction assertions, and watchtowers.

Header Chain Method IDs

These constants represent the method IDs for different network header chains, such as Mainnet, Testnet4, Signet, and Regtest.

Constants

FIRST_FIVE_OUTPUTS

The number of kickoff outputs before the first assert utxo.

MAINNET_HEADER_CHAIN_METHOD_ID

MAX_NUMBER_OF_WATCHTOWERS

The theoretical maximum number of watchtowers that can be used in the Clementine protocol.

NUMBER_OF_ASSERT_TXS

The number of assertion transactions that a challenged operator should send.

REGTEST_HEADER_CHAIN_METHOD_ID

SIGNET_HEADER_CHAIN_METHOD_ID

TESTNET4_HEADER_CHAIN_METHOD_ID